requirements.txt
pyautogen>=0.2.0
numpy
pandas
colorama
